fixing delegate directly in am
